+++ CHANGELOG.md
@@ -20,5 +20,6 @@ and this project adheres to
 - 🔒(authentication) make the logout view POST-only
 - 👽(authentication) handle Social Auth's login views requiring POST requests
 - 🦖(OIDC Provider) make introspection endpoint fall back to PSA backends
+- ✨(OIDC Provider) relay RP-Initiated logout confirmation to the frontend
 
 [unreleased]: https://github.com/suitenumerique/accounts/compare/main